create dashboard admin

+++ datasdngis/routes/web.php
@@ -4,10 +4,6 @@ use App\Http\Controllers\AuthController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\SdnController;
 
-// Route publik
-Route::get('/', [SdnController::class, 'index'])->name('sdn.index');
-Route::get('/sdn/{slug}', [SdnController::class, 'show'])->name('sdn.show');
-
 // Route autentikasi
 Route::middleware(['guest'])->group(function () {
     Route::get('/login', [AuthController::class, 'index'])->name('login.page');
@@ -17,9 +13,15 @@ Route::middleware(['guest'])->group(function () {
 // Route yang memerlukan autentikasi
 Route::middleware(['auth'])->group(function () {
     Route::get('/logout', [AuthController::class, 'logout'])->name('logout');
-    Route::get('/sdn/create', [SdnController::class, 'create'])->name('sdn.create');
-    Route::post('/sdn', [SdnController::class, 'store'])->name('sdn.store');
-    Route::get('/sdn/{slug}/edit', [SdnController::class, 'edit'])->name('sdn.edit');
-    Route::put('/sdn/{slug}', [SdnController::class, 'update'])->name('sdn.update');
-    Route::delete('/sdn/{slug}', [SdnController::class, 'destroy'])->name('sdn.destroy');
+    Route::get('/admin/dashboard', [SdnController::class, 'dashboard'])->name('admin.dashboard');
+
+    // CRUD SDN
+    Route::get('/admin/sdn/create', [SdnController::class, 'create'])->name('sdn.create');
+    Route::post('/admin/sdn', [SdnController::class, 'store'])->name('sdn.store');
+    Route::get('/admin/sdn/{slug}/edit', [SdnController::class, 'edit'])->name('sdn.edit');
+    Route::put('/admin/sdn/{slug}', [SdnController::class, 'update'])->name('sdn.update');
+    Route::delete('/admin/sdn/{slug}', [SdnController::class, 'destroy'])->name('sdn.destroy');
 });
+
+// Route publik
+Route::get('/', [SdnController::class, 'index'])->name('sdn.index');
